+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Changes in SimDesign 2.27
 
+- `descript()` provides better tracking of variable name information
+
 - `SimCollect()` works correctly with count data now and avoids 
   Jensen inequality behavior when weighted averages are inappropriate 
   (reported by Michael Troung)